Create opendiscovery_odscreen
Automated computational docking of large libraries of chemical compounds to a protein can aid in pharmaceutical drug design and gives scientists with basic computer experience a tool to help plan wet laboratory investigations when exploring the combination of chemical and pharmacological spaces. The use of open source tools to develop and select ligands for subsequent screening is outlined. This is a protocol leveraging the power of Open Babel and AutoDock Vina to perform file conversion, minimization, and docking implemented as a Python script